requirements.txt
pandas
requests
joblib
xgboost
python-dotenv
